v2-alpha-6 - Added Disqus Comment section in footer - Some File and Components management
