+++ package.json
@@ -1,12 +1,13 @@
 {
   "name": "jntuk-instant",
-  "version": "2.0.0-alpha-5.2",
+  "version": "2.0.0-alpha-6",
   "private": true,
   "homepage": "https://jntuk-instant.github.io/The-React-Port/",
   "dependencies": {
     "@testing-library/jest-dom": "^5.11.9",
     "@testing-library/react": "^11.2.3",
     "@testing-library/user-event": "^12.6.0",
+    "query-string": "^4.3.4",
     "react": "^17.0.1",
     "react-dom": "^17.0.1",
     "react-router-dom": "^5.2.0",
