+++ public/index.html
@@ -14,5 +14,25 @@
   <body>
     <div id="root"></div>
     <noscript>You need to enable JavaScript to view this Website.</noscript>
+
+    <script>
+      /**
+       *  RECOMMENDED CONFIGURATION VARIABLES: EDIT AND UNCOMMENT THE SECTION BELOW TO INSERT DYNAMIC VALUES FROM YOUR PLATFORM OR CMS.
+       *  LEARN WHY DEFINING THESE VARIABLES IS IMPORTANT: https://disqus.com/admin/universalcode/#configuration-variables*/
+
+      var disqus_config = function () {
+        this.page.url = "https://jntuk-instant.github.io/The-React-Port/"; // Replace PAGE_URL with your page's canonical URL variable
+        this.page.identifier = "123456789"; // Replace PAGE_IDENTIFIER with your page's unique identifier variable
+      };
+
+      (function () {
+        // DON'T EDIT BELOW THIS LINE
+        var d = document,
+          s = d.createElement("script");
+        s.src = "https://https-jntuk-instant-github-io.disqus.com/embed.js";
+        s.setAttribute("data-timestamp", +new Date());
+        (d.head || d.body).appendChild(s);
+      })();
+    </script>
   </body>
 </html>
