+++ src/components/footer.js
@@ -1,0 +1,15 @@
+import "../styles/footer.css";
+
+
+function Footer() {
+  return (
+    <footer>
+      <div id="disqus_thread"></div>
+      <noscript>
+        Please enable JavaScript to view the <a href="https://disqus.com/?ref_noscript">comments powered by Disqus.</a>
+      </noscript>
+    </footer>
+  );
+}
+
+export default Footer;
