--- src/pages/about.js
@@ -1,12 +1,0 @@
-//import react from "react";
-
-function About() {
-  return (
-    <div>
-      <p>Made with React and react-router-dom</p>
-      <p>v2-alpha</p>
-    </div>
-  );
-}
-
-export default About;
