+++ src/styles/footer.css
@@ -1,0 +1,8 @@
+footer {
+  width: 100%;
+  min-height: 100px;
+  background-color: hsla(48, 100%, 93%, 1);
+  border-radius: 0 0 30px 30px;
+  box-sizing: border-box;
+  padding: 7px;
+}
